+++ app/helpers/shoppe/application_helper.rb
@@ -1,5 +1,9 @@
 module Shoppe
   module ApplicationHelper
+    
+    def navigation_manager_link(item)
+      link_to item.description, item.url, item.link_options.merge(:class => item.active?(self) ? 'active' : 'inactive')
+    end
   
     def status_tag(status)
       content_tag :span, status, :class => "status-tag #{status}"